Add ESG metrics page and improve benchmarking
