+++ src/env.d.ts
@@ -3,6 +3,7 @@
 interface ImportMetaEnv {
   readonly VITE_SUPABASE_URL: string;
   readonly VITE_SUPABASE_ANON_KEY: string;
+  readonly VITE_STRIPE_PUBLISHABLE_KEY?: string;
   readonly VITE_CLERK_PUBLISHABLE_KEY?: string;
   readonly VITE_CLERK_SECRET_KEY?: string;
   readonly BASE_URL?: string;
