+++ src/pages/HomePage.tsx
@@ -1,0 +1,5 @@
+import Home from "../components/home";
+
+export default function HomePage() {
+  return <Home />;
+}
